:house: Change camel case files to snake case
