+++ src/commands/broadcast_signature.js
