+++ src/commands/create_transaction_register_delegate.js
