+++ src/commands/create_transaction_register_second_passphrase.js
