+++ src/commands/show_copyright.js
