+++ src/commands/show_warranty.js
