+++ src/utils/crypto_module.js
