+++ test/_global_hooks.js
