+++ test/specs/commands/broadcast_signature.js
