+++ test/specs/commands/broadcast_transaction.js
