+++ test/specs/commands/create_transaction_register_multisignature_account.js
