+++ test/specs/commands/create_transaction_transfer.js
