+++ test/specs/commands/decrypt_message.js
