+++ test/specs/commands/decrypt_passphrase.js
