+++ test/specs/commands/encrypt_message.js
