+++ test/specs/commands/encrypt_passphrase.js
