+++ test/specs/commands/show_copyright.js
