+++ test/specs/commands/show_warranty.js
