+++ test/specs/commands/sign_message.js
