+++ test/steps/setup.js
@@ -19,7 +19,7 @@ import readline from 'readline';
 import lockfile from 'lockfile';
 import lisk from 'lisk-js';
 import liskAPIInstance from '../../src/utils/api';
-import cryptoInstance from '../../src/utils/cryptoModule';
+import cryptoInstance from '../../src/utils/crypto_module';
 import * as fsUtils from '../../src/utils/fs';
 import * as helpers from '../../src/utils/helpers';
 import * as input from '../../src/utils/input';
